requirements.txt
python-dotenv
gradio
langchain
langchain-community
langchain-huggingface
langchain-experimental
transformers
torch
sentence-transformers
huggingface_hub
pypdf
faiss-cpu

